Change mv_t to int32_t because of possible overflow in large videos

+++ src/global.h
@@ -126,7 +126,7 @@
 
 typedef int16_t coeff_t;
 
-typedef int16_t mv_t;
+typedef int32_t mv_t;
 
 //#define VERBOSE 1
 //#define UVG_DEBUG_PRINT_CABAC 1
